Add AModContent and SVMod classes
A lot of the previous work was just trying to get things working and figuring out what I need to do. Now I'm starting to focus more on streamlining the code and also setting it up to be a proper mod manager in the future. I'm likely going to move more methods into the ModContent and SVMod classes, as well as add fields that track what mods everything comes from so that you can toggle mods on/off in-game.