Add localized strings
All AModContent and ContentModification classes now have localized dictionary versions of their string values. The keys are the locale codes of the target language (can be found in the in-game language selector). The original displayName/description values are treated as defaults that only apply if no localization is specified.

Removed AModPilot's true ending dialogue properties, and replaced them with the localization dictionary. (The game already provides a default for the true ending dialogue, so I don't want modders overriding them unless it's with properly localized text. Meanwhile, for everything else, the "No translation" text would be worse than the original untranslated text since it wouldn't include ANY form of the content's name, only its number id, so I'm fine with using a global default for them.)